% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_pixel}
\alias{mouse_pixel}
\title{Mouse Liver Final Pixel Assignment}
\format{
A data frame with 60170 rows and 15 variables:
\describe{
\item{topK}{Final pixel cell type assignment based on FICTURE model}
\item{X}{pixel X coordinate.}
\item{Y}{pixel Y coordinate.}
}
}
\source{
Generated from the slda_decoder function.
}
\usage{
data(mouse_updated_anchor)
}
\description{
A dataset containing final class assignment for each pixel point.
}
\keyword{datasets}
