% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_transcripts}
\alias{mouse_transcripts}
\title{Mouse Transcript Dataset}
\format{
A data frame with 5 rows and 3 variables:
}
\source{
Generated from a CSV file and saved as R data.
}
\usage{
mouse_transcripts
}
\description{
A dataset containing sample names and scores.
}
\keyword{datasets}
