% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda_decoder.R
\name{slda_decoder}
\alias{slda_decoder}
\title{Decode the spatial transcriptomics data using the anchor-based method}
\usage{
slda_decoder(transcripts, anchor)
}
\arguments{
\item{transcripts}{A data frame containing pixel level spatial transcriptomics data.}

\item{anchor}{A data frame containing initial anchor level assignments and probabilities.}
}
\value{
Updated anchor level data after decoding.
}
\description{
This function decodes the spatial transcriptomics data using the anchor-based method.
}
\examples{
slda_decoder(transcripts, anchor)
}
