% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_anchors}
\alias{mouse_anchors}
\title{Mouse Anchors Initial Dataset}
\format{
A data frame with 31322 rows and 18 variables:
\describe{
\item{unit}{Anchor Idenitifier}
\item{topK}{Initial cell type assignment based on FICTURE model}
\item{X}{Anchor X coordinate.}
\item{Y}{Anchor Y coordinate.}
\item{Count}{# of genes is observed within the anchor space}
}
}
\source{
Sourced from the original python FICTURE example and Vizgen MERFISH Mouse Liver Map.
}
\usage{
data(mouse_updated_anchor)
}
\description{
A dataset containing sample names and scores.
}
\keyword{datasets}
