% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_output}
\alias{mouse_output}
\title{Mouse Liver Final Anchor Assignment}
\format{
A data frame with 675 rows and 15 variables:
\describe{
\item{topK}{Initial cell type assignment based on FICTURE model}
\item{X}{Anchor X coordinate.}
\item{Y}{Anchor Y coordinate.}
}
}
\source{
Generated from the slda_decoder function.
}
\usage{
mouse_output
}
\description{
A dataset containing final class assignment for each anchor point.
}
\keyword{datasets}
