% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_transcripts}
\alias{mouse_transcripts}
\title{Mouse Liver Gene Transcripts}
\format{
A data frame with 2,462,371 rows and 5 variables:
\describe{
\item{gene}{Gene name for each observation}
\item{X}{Pixel X coordinate.}
\item{Y}{Pixel Y coordinate.}
\item{Count}{# of times gene is observed at given coordinate}
}
}
\source{
Sourced from the original python FICTURE example and Vizgen MERFISH Mouse Liver Map.
}
\usage{
data(mouse_updated_anchor)
}
\description{
A dataset containing gene observations labeled with X and Y coordinates and observation counts
}
\keyword{datasets}
